Add mouse events to components
In React, you want to use state sparingly. For the most part, everything
should go through this.props and only use state to trigger rerenders.

The main problem with using state is that the state tends to stick around
even when you no longer want it. For example, when rendering the children,
if we used the children array from `this.state.node`, then when deleting
nodes, the children would merely shift positions. So even though one node
might have a lot of children and we delete a node without any children,
the lad with children would get shifted to the old position which didn't
have any children, and as a consequence it would no longer have its
children displayed.

Small shit like that compounds into larger, harder-to-figure-out bugs.